server/requirements.txt
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
docusign-esign==5.3.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
nose==1.3.7
pycparser==2.22
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==75.5.0
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
requests==2.*
Flask-Session>=0.5.0